demlo-backend/backend/requirements.txt
asgiref==3.4.1
bcrypt==3.2.0
cryptography==3.4.7
Django==3.2.5
djangorestframework==3.12.4
paramiko==2.7.2
pycparser==2.20
PyNaCl==1.4.0
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
typing-extensions==3.10.0.0
djangorestframework-simplejwt==5.0.0
pycryptodomex
psycopg2-binary==2.9.3
django-environ==0.8.1